Optimize code. Optimize core tests. Improve code quality. (#5749)
* Optimize core tests [1] --filter=[core]

* Optimize core tests [2], Optimize server code

* Optimize server code [2]

* Fix tests

* Fix tests

* Added more tests

* Fix

* Fix 2

* Fix 3 --filter=[core]

* Fix 4 --filter=[core]

* Fix 5 --filter=[core]

* Added System::get_dns_cache_hit_ratio(), fix tests --filter=[core]

* Optimize logger impl, add Logger::set_stream()  --filter=[core]

* Fix tests [6] --filter=[core]

* Fix tests [7] --filter=[core]

* Fix tests [8] --filter=[core]

* Optimize tests [9] --filter=[core]

+++ src/coroutine/socket.cc
@@ -386,7 +386,7 @@ bool Socket::http_proxy_handshake() {
         return false;
     }
 
-    swoole_trace_log(SW_TRACE_HTTP_CLIENT, "proxy response: <<EOF\n%.*sEOF", n, recv_buffer->str);
+    swoole_trace_log(SW_TRACE_HTTP_CLIENT, "proxy response: <<EOF\n%.*sEOF", (int) n, recv_buffer->str);
 
     if (!http_proxy->handshake(recv_buffer)) {
         set_err(SW_ERROR_HTTP_PROXY_BAD_RESPONSE,
